\name{sampledata}
\alias{sampledata}
\title{sample input data}
\description{sample input data}
\usage{data(sampledata)}
\value{sample input data}
\format{comma separated value file, containing five columns: 
        i_exp (experiment index), i_node (node_index), 
        outcome (-1/0/1), value (cost for that outcome), 
        is_perturbation (0 or 1)}
